+++ common/db/src/test/java/is/codion/common/db/exception/DatabaseExceptionTest.java
@@ -21,6 +21,7 @@ package is.codion.common.db.exception;
 import org.junit.jupiter.api.Test;
 
 import java.sql.SQLException;
+import java.util.Optional;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
@@ -33,10 +34,10 @@ public class DatabaseExceptionTest {
 			throw new DatabaseException("hello");
 		});
 		DatabaseException dbException = new DatabaseException("hello", "statement");
-		assertEquals("statement", dbException.statement());
+		assertEquals(Optional.of("statement"), dbException.statement());
 		dbException = new DatabaseException(new SQLException(), "message", "statement");
 		assertEquals("message", dbException.getMessage());
-		assertEquals("statement", dbException.statement());
+		assertEquals(Optional.of("statement"), dbException.statement());
 		dbException = new DatabaseException(null, "test", "stmt");
 		assertEquals(-1, dbException.errorCode());
 	}